Revise entrypoint renaming interface.
Changed the interface from `IEntryPoint::getRenamedEntryPoint` to `IComponentType::renameEntryPoint`.

The underlying implementation creates a `RenamedEntryPointComponentType` wrapper object around the base entry-point.

This new implementation allows the user to specify entry point renaming on an IComponentType that isn't just a `EntryPoint`, but also on `SpecializedComponentType` or `CompositeComponentType` as long as the component defines a single entry point.