Fix folding of no-arg constructs in SCCP pass
The SCCP optimization pass was recently updated to include some actual
constant folding, but the case for handling `kIROp_Construct` is
mistakenly assuming that any `construct` instruction of a basic
scalar type *must* have exaclty one operand. We have multiple places
in the code that form zero-argument `construct` instructions with
the understanding that these will represent "default" values, and if
one of those ends up in SCCP it causes a crash.

This change adds simple logic to make the folding more robust by
checking the operand count of `construct` instructions before
processing them, and it also adds the logic to fold the zero-operand
case, since that should be a simple one.