Move the API headers under an include directory
When the project was smaller in size and scope, it made a lot of sense
to have the (single) public API header be at the top level of the
repository where users could easily find and explore it. As the
project has grown, so too has the size and complexity of the headers
involved. We now have multiple headers (even ignoring GFX), and the
`slang.h` header is much less readable as a result of accumulated
boilerplate at the top.

This change moves the headers under `include/slang/`. A subsequent
change will add "forwarding" headers at the root of the project for
`slang.h` and `slang-gfx.h`. Those changes were explicitly *not*
included in this change, in the hopes that source-control systems
can detect that this change is *moving* the files.