Add design proposal for basic interfaces.
This change adds a `docs/proposals/` directory to give us a place to iterate on proposals for language/compiler features that are a bit too large to handle as issues.
The first proposal I'm checking in here is a sketch of what our built-in numeric interfaces might want to look like.

I do not propose that the design in the document is perfect or even *finished*.
The goal here is to get things checked into source control to foster subsequent discussion and iteration.