Enable -Wunused-variable
We tend to do `if(auto unused = as<Foo>(bar))` a lot, where `unused` is
a descriptive name. In most of these cases this changes these to be
`const auto` which gcc does not warn about.

Fixed one bug where we weren't passing IRCFGNormalizationPass to
normalizeCFG in prepareFuncForBackwardDiff (for the sink).