Add support for texture footprint queries
This change adds operations to the Slang standard library that allow applications to query the texels that would be accessed by a texture sampling operation.
These new operations are implementated for Vulkan via the `GL_NV_shader_texure_footprint` extension, and for D3D via NVAPI.

The main challenge in implementing this is that the NVAPI functions require the space/register of the texture and sampler being used for the query to be passed in as integers.
In order to support this, the change adds new IR instructions that query the binding information of a resource, and a specialization pass that lowers those instructions by modifying the program to pass space/register information through the call chain.

The tests that have been added confirm that valid SPIR-V is being generated for this new API, but only validates the HLSL output for the D3D/NVAPI case.