Fix nonuniform decoration on direct-to-spirv backend path. (#3338)
By updating the hlsl meta intrinsic for spirv path. Decorating the index that is
input to NonUniformResourceIndex was getting removed by spirv-opt, when
inlining as the decoration was applied to the function parameter.
Thus the implementation was modified by introducing OpCopyObject that
copies the object inside the function body of the meta builtin
"NonUniformResourceIndex" and the decoration is applied to the result of
OpCopyObject. With this approach SPIRV-OPt Inliner propagates decoration correctly.