Fix incorrect behavior of operator%
Fixes #1059.

This change fixes the incorrect translation of "operator%" from
HLSL to SPIRV.

The issue stems from the fact that the behavior of "operator%" in
GLSL differs from that in HLSL. In HLSL it behaves as "remainder"
where as it behaves as "modulus" in GLSL.

We have been using SpvOpFMod for operator% when Slang compiles
from HLSL to SPRIV, which is incorrect.  This change switches it
to SpvOpFRem.

The tests are slightly modified to reveal any potential issues.