Fix incorrect behavior of operator% (#3470)
* Fix incorrect behavior of operator%

Fixes #1059.

This change fixes the incorrect translation of "operator%" from
HLSL to SPIRV.

The issue stems from the fact that the behavior of "operator%" in
GLSL differs from that in HLSL. In HLSL it behaves as "remainder"
where as it behaves as "modulus" in GLSL.

We have been using SpvOpFMod for operator% when Slang compiles
from HLSL to SPRIV, which is incorrect.  This change switches it
to SpvOpFRem.

The tests are slightly modified to reveal any potential issues.

* Change output type of test/compute/frem

For testing the operator%, we are using "int" as the output type of the
test, "test/compute/frem.slang".

Since the operands are in float type, it is more preferable to have a
float type as the resulting type. This can be done with an option,
"-output-using-type".

---------

Co-authored-by: Yong He <yonghe@outlook.com>