Preserve the precision of a small value near zero
Fixes an issue #3490

When a very small value is emitted, it was snapping to a zero value.
This change preserves the precision more to keep the value from being a
zero value.

When "ostringstream" was printing a floating point value, we set
"presicion" that indicates the number of digits in decimal. We have
been using a fixed value, 20, which limits the available value near
zero to be 0.00000000000000000001. When a given value is smaller, the
mantissa part was truncated and it became a zero.

This change calculates how much precision is needed based on the given
value and the type. The mantissa bits of "float" type can represent
upto 8 digits and that of "double" type can represent upto 16 digits
in decimal. The exponential part of "float" can represent upto 38
digits and that of "double" can represent upto 308 digits.